aggregator tests cleaned and fixed
